Refactor auto reply code
Thanks to @Jupeyy for helping with nullptr buggos

Go follow him on github https://github.com/Jupeyy he is top 10
underrated teeworlds githubbers!!!
