+++ src/game/client/components/chillerbot/chathelper/check_team.cpp
@@ -1,0 +1,26 @@
+// chillerbot-ux reply to ping
+
+#include "engine/shared/protocol.h"
+#include <engine/client/notifications.h>
+#include <engine/config.h>
+#include <engine/console.h>
+#include <engine/graphics.h>
+#include <engine/keys.h>
+#include <engine/textrender.h>
+#include <game/client/animstate.h>
+#include <game/client/components/camera.h>
+#include <game/client/components/chat.h>
+#include <game/client/components/chillerbot/chathelper.h>
+#include <game/client/components/chillerbot/version.h>
+#include <game/client/components/controls.h>
+#include <game/client/components/menus.h>
+#include <game/client/components/voting.h>
+#include <game/client/gameclient.h>
+#include <game/client/prediction/entities/character.h>
+#include <game/client/race.h>
+#include <game/client/render.h>
+#include <game/generated/client_data.h>
+#include <game/generated/protocol.h>
+#include <game/version.h>
+
+#include "replytoping.h"
