+++ src/game/client/components/chillerbot/chathelper/replytoping.cpp
@@ -38,247 +38,6 @@ CReplyToPing::CReplyToPing(CChatHelper *pChatHelper, const char *pMessageAuthor,
 	m_SizeOfResponse = SizeOfResponse;
 }
 
-bool CReplyToPing::WhyWar(const char *pVictim, bool IsCheck)
-{
-	if(!pVictim)
-		return false;
-
-	bool HasWar = true;
-	// aVictim also has to hold the full own name to match the chop off
-	char aVictim[MAX_NAME_LENGTH + 3 + MAX_NAME_LENGTH];
-	str_copy(aVictim, pVictim, sizeof(aVictim));
-	if(!ChatHelper()->GameClient()->m_WarList.IsWarlist(aVictim) && !ChatHelper()->GameClient()->m_WarList.IsTraitorlist(aVictim))
-	{
-		HasWar = false;
-		while(str_endswith(aVictim, "?")) // cut off the question marks from the victim name
-			aVictim[str_length(aVictim) - 1] = '\0';
-		while(str_endswith(aVictim, " ")) // cut off spaces from victim name 'why war foo ?' -> 'foo ?' -> 'foo ' -> 'foo'
-			aVictim[str_length(aVictim) - 1] = '\0';
-		// cut off own name from the victime name if question in this format "why do you war foo (your name)"
-		char aOwnName[MAX_NAME_LENGTH + 3];
-		// main tee
-		str_format(aOwnName, sizeof(aOwnName), " %s", ChatHelper()->GameClient()->m_aClients[ChatHelper()->GameClient()->m_LocalIDs[0]].m_aName);
-		if(str_endswith_nocase(aVictim, aOwnName))
-			aVictim[str_length(aVictim) - str_length(aOwnName)] = '\0';
-		if(ChatHelper()->GameClient()->Client()->DummyConnected())
-		{
-			str_format(aOwnName, sizeof(aOwnName), " %s", ChatHelper()->GameClient()->m_aClients[ChatHelper()->GameClient()->m_LocalIDs[1]].m_aName);
-			if(str_endswith_nocase(aVictim, aOwnName))
-				aVictim[str_length(aVictim) - str_length(aOwnName)] = '\0';
-		}
-
-		// cut off descriptions like this
-		// why do you block foo he is new here!
-		// why do you block foo she is my friend!!
-		for(int i = 0; i < str_length(aVictim); i++)
-		{
-			// c++ be like...
-			if(i < 2)
-				continue;
-			if(aVictim[i - 1] != ' ')
-				continue;
-			if((aVictim[i] != 'h' || !aVictim[i + 1] || aVictim[i + 1] != 'e' || !aVictim[i + 2] || aVictim[i + 2] != ' ') &&
-				(aVictim[i] != 's' || !aVictim[i + 1] || aVictim[i + 1] != 'h' || !aVictim[i + 2] || aVictim[i + 2] != 'e' || !aVictim[i + 3] || aVictim[i + 3] != ' '))
-				continue;
-
-			aVictim[i - 1] = '\0';
-			break;
-		}
-
-		// do not kill my friend foo
-		const char *pFriend = NULL;
-		if((pFriend = str_find_nocase(aVictim, " friend ")))
-			pFriend += str_length(" friend ");
-		else if((pFriend = str_find_nocase(aVictim, " frint ")))
-			pFriend += str_length(" frint ");
-		else if((pFriend = str_find_nocase(aVictim, " mate ")))
-			pFriend += str_length(" mate ");
-		else if((pFriend = str_find_nocase(aVictim, " bff ")))
-			pFriend += str_length(" bff ");
-		else if((pFriend = str_find_nocase(aVictim, " girlfriend ")))
-			pFriend += str_length(" girlfriend ");
-		else if((pFriend = str_find_nocase(aVictim, " boyfriend ")))
-			pFriend += str_length(" boyfriend ");
-		else if((pFriend = str_find_nocase(aVictim, " dog ")))
-			pFriend += str_length(" dog ");
-		else if((pFriend = str_find_nocase(aVictim, " gf ")))
-			pFriend += str_length(" gf ");
-		else if((pFriend = str_find_nocase(aVictim, " bf ")))
-			pFriend += str_length(" bf ");
-
-		if(pFriend)
-			str_copy(aVictim, pFriend, sizeof(aVictim));
-	}
-
-	char aWarReason[128];
-	if(HasWar || ChatHelper()->GameClient()->m_WarList.IsWarlist(aVictim) || ChatHelper()->GameClient()->m_WarList.IsTraitorlist(aVictim))
-	{
-		ChatHelper()->GameClient()->m_WarList.GetWarReason(aVictim, aWarReason, sizeof(aWarReason));
-		if(aWarReason[0])
-			str_format(m_pResponse, m_SizeOfResponse, "%s: %s has war because: %s", m_pMessageAuthor, aVictim, aWarReason);
-		else
-			str_format(m_pResponse, m_SizeOfResponse, "%s: the name %s is on my warlist.", m_pMessageAuthor, aVictim);
-		return true;
-	}
-	for(int i = 0; i < MAX_CLIENTS; i++)
-	{
-		auto &Client = ChatHelper()->GameClient()->m_aClients[i];
-		if(!Client.m_Active)
-			continue;
-		if(str_comp(Client.m_aName, aVictim))
-			continue;
-
-		if(ChatHelper()->GameClient()->m_WarList.IsWarClanlist(Client.m_aClan))
-		{
-			str_format(m_pResponse, m_SizeOfResponse, "%s i war %s because his clan %s is on my warlist.", m_pMessageAuthor, aVictim, Client.m_aClan);
-			return true;
-		}
-		if(ChatHelper()->GameClient()->m_WarList.IsWarClanmate(i))
-		{
-			str_format(m_pResponse, m_SizeOfResponse, "%s i might kill %s because I war member from his clan %s", m_pMessageAuthor, aVictim, Client.m_aClan);
-			return true;
-		}
-	}
-	if(IsCheck && str_comp_nocase(aVictim, "me"))
-	{
-		str_format(m_pResponse, m_SizeOfResponse, "%s: '%s' is not on my warlist.", m_pMessageAuthor, aVictim);
-		return true;
-	}
-	return false;
-}
-
-int CReplyToPing::GetSuffixLen(const char *pStr, const char *pSuffix)
-{
-	if(str_endswith(pStr, pSuffix))
-		return str_length(pSuffix);
-	return 0;
-}
-
-void CReplyToPing::StripSpacesAndPunctuationAndOwnName(const char *pStr, char *pStripped, int SizeOfStripped)
-{
-	dbg_assert(SizeOfStripped < 512, "too big to strip");
-	if(pStr == pStripped)
-	{
-		pStripped[0] = '\0';
-		return;
-	}
-	if(!pStr)
-	{
-		pStripped[0] = '\0';
-		return;
-	}
-	if(!pStripped)
-	{
-		pStripped[0] = '\0';
-		return;
-	}
-	if(SizeOfStripped < 1)
-	{
-		pStripped[0] = '\0';
-		return;
-	}
-	char aBuf[512];
-	str_copy(aBuf, pStr, sizeof(aBuf));
-	while(str_endswith(aBuf, "?")) // cut off the question marks
-		aBuf[str_length(aBuf) - 1] = '\0';
-	while(str_endswith(aBuf, " ")) // cut off spaces
-		aBuf[str_length(aBuf) - 1] = '\0';
-	int Offset = 0;
-	const char *pName = ChatHelper()->GameClient()->m_aClients[ChatHelper()->GameClient()->m_LocalIDs[0]].m_aName;
-	const char *pDummyName = ChatHelper()->GameClient()->m_aClients[ChatHelper()->GameClient()->m_LocalIDs[1]].m_aName;
-	char aName[128];
-	str_format(aName, sizeof(aName), "%s: ", pName);
-	if(!Offset && str_startswith(pStr, aName))
-		Offset = str_length(aName);
-	str_format(aName, sizeof(aName), "%s ", pName);
-	if(!Offset && str_startswith(pStr, aName))
-		Offset = str_length(aName);
-	str_format(aName, sizeof(aName), "%s", pName);
-	if(!Offset && str_startswith(pStr, aName))
-		Offset = str_length(aName);
-	if(ChatHelper()->GameClient()->Client()->DummyConnected())
-	{
-		str_format(aName, sizeof(aName), "%s: ", pDummyName);
-		if(!Offset && str_startswith(pStr, aName))
-			Offset = str_length(aName);
-		str_format(aName, sizeof(aName), "%s ", pDummyName);
-		if(!Offset && str_startswith(pStr, aName))
-			Offset = str_length(aName);
-		str_format(aName, sizeof(aName), "%s", pDummyName);
-		if(!Offset && str_startswith(pStr, aName))
-			Offset = str_length(aName);
-	}
-	if(Offset >= str_length(aBuf))
-		pStripped[0] = '\0';
-	else
-		str_copy(pStripped, aBuf + Offset, SizeOfStripped);
-}
-
-bool CReplyToPing::IsEmptyStr(const char *pStr)
-{
-	if(!pStr)
-		return true;
-	for(int i = 0; pStr[i] != '\0'; i++)
-		if(pStr[i] != ' ' && pStr[i] != 0x7)
-			return false;
-	return true;
-}
-
-int CReplyToPing::IsWarCheckSuffix(const char *pStr)
-{
-	const char aaPrefixes[][16] = {" in", " on", " at", " is"};
-	const char aaYous[][16] = {" u", " you", " your", " ur", " deiner", " deinen", " ju", " ti", " jour", " jur", " tu"};
-	const char aaWarlists[][64] = {
-		" warlist",
-		" war list",
-		" friendlist",
-		" friend list",
-		" team list",
-		" peace list",
-		" enemy list",
-		" enemies",
-		" enemys",
-		" frint list",
-		" frintlist",
-		" frentlist",
-		" frent list",
-		" frenlist",
-		" fren list",
-		" friends",
-		" frints",
-		" frintss",
-		" frents",
-		" friend",
-		" frint",
-		" frent",
-		" fren",
-		" good",
-		" war",
-		" kill",
-		" bad"};
-	char aOnYourWarlist[128];
-	int ChopEnding = 0;
-	for(const auto &aPrefix : aaPrefixes)
-	{
-		for(const auto &aWarlist : aaWarlists)
-		{
-			str_format(aOnYourWarlist, sizeof(aOnYourWarlist), "%s%s", aPrefix, aWarlist);
-			dbg_msg("chiller", "check='%s'", aOnYourWarlist);
-			ChopEnding = GetSuffixLen(pStr, aOnYourWarlist);
-			if(ChopEnding)
-				return ChopEnding;
-			for(const auto &aYou : aaYous)
-			{
-				str_format(aOnYourWarlist, sizeof(aOnYourWarlist), "%s%s%s", aPrefix, aYou, aWarlist);
-				dbg_msg("chiller", "check='%s'", aOnYourWarlist);
-				if((ChopEnding = GetSuffixLen(pStr, aOnYourWarlist)))
-					return ChopEnding;
-			}
-		}
-	}
-	return ChopEnding;
-}
-
 bool CReplyToPing::Reply()
 {
 	if(!m_pResponse)
@@ -338,589 +97,29 @@ bool CReplyToPing::Reply()
 			return true;
 		}
 	}
-	// check if a player has war or not
-	const char *pDoYou = nullptr;
-	if(!pDoYou && (pDoYou = str_find_nocase(m_pMessage, "you war ")))
-		pDoYou = pDoYou + str_length("you war ");
-	if(!pDoYou && (pDoYou = str_find_nocase(m_pMessage, "you in war with ")))
-		pDoYou = pDoYou + str_length("you in war with ");
-	// "hast du war mit"
-	// "hast du eig war mit"
-	// "hast du eigentlich war mit"
-	// "hast du überhaupt war mit"
-	// "hast du einen war mit"
-	if(!pDoYou && (pDoYou = LangParser().StrFindOrder(m_pMessage, 2, "hast du ", "war mit ")))
-		pDoYou = pDoYou + str_length("war mit ");
-	if(pDoYou)
-		if(WhyWar(pDoYou, true))
-			return true;
-
-	// check war reason for others
-	const char *pWhy = str_find_nocase(m_pMessage, "why has ");
-	if(pWhy)
-		pWhy = pWhy + str_length("why has ");
-	if(!pWhy)
-		if((pWhy = str_find_nocase(m_pMessage, "why")))
-			pWhy = pWhy + str_length("why");
-	if(!pWhy)
-		if((pWhy = str_find_nocase(m_pMessage, "warum hat ")))
-			pWhy = pWhy + str_length("warum hat ");
-	if(!pWhy)
-		if((pWhy = str_find_nocase(m_pMessage, "warum")))
-			pWhy = pWhy + str_length("warum");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "stop");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "do not");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "don't");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "do u");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "do you");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "is u");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "is you");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "are u");
-	if(!pWhy)
-		pWhy = str_find_nocase(m_pMessage, "are you");
-	if(pWhy)
-	{
-		const char *pKill = NULL;
-		pKill = str_find_nocase(pWhy, "kill "); // why do you kill foo?
-		if(pKill)
-			pKill = pKill + str_length("kill ");
-		else if((pKill = str_find_nocase(pWhy, "kil "))) // why kil foo?
-			pKill = pKill + str_length("kil ");
-		else if((pKill = str_find_nocase(pWhy, "killed "))) // why killed foo?
-			pKill = pKill + str_length("killed ");
-		else if((pKill = str_find_nocase(pWhy, "block "))) // why do you block foo?
-			pKill = pKill + str_length("block ");
-		else if((pKill = str_find_nocase(pWhy, "blocked "))) // why do you blocked foo?
-			pKill = pKill + str_length("blocked ");
-		else if((pKill = str_find_nocase(pWhy, "help "))) // why no help foo?
-			pKill = pKill + str_length("help ");
-		else if((pKill = str_find_nocase(pWhy, "war with "))) // why do you have war with foo?
-			pKill = pKill + str_length("war with ");
-		else if((pKill = str_find_nocase(pWhy, "war "))) // why war foo?
-			pKill = pKill + str_length("war ");
-		else if((pKill = str_find_nocase(pWhy, "wared "))) // why wared foo?
-			pKill = pKill + str_length("wared ");
-		else if((pKill = str_find_nocase(pWhy, "waring "))) // why waring foo?
-			pKill = pKill + str_length("waring ");
-		else if((pKill = str_find_nocase(pWhy, "bully "))) // why bully foo?
-			pKill = pKill + str_length("bully ");
-		else if((pKill = str_find_nocase(pWhy, "bullying "))) // why bullying foo?
-			pKill = pKill + str_length("bullying ");
-		else if((pKill = str_find_nocase(pWhy, "bullied "))) // why bullied foo?
-			pKill = pKill + str_length("bullied ");
-		else if((pKill = str_find_nocase(pWhy, "freeze "))) // why freeze foo?
-			pKill = pKill + str_length("freeze ");
-		else if((pKill = str_find_nocase(pWhy, "warlist "))) // is warlist foo?
-			pKill = pKill + str_length("warlist ");
-		else if((pKill = str_find_nocase(pWhy, "enemi "))) // is enemi foo?
-			pKill = pKill + str_length("enemi ");
-		else if((pKill = str_find_nocase(pWhy, "enemy "))) // is enemy foo?
-			pKill = pKill + str_length("enemy ");
-
-		if(WhyWar(pKill))
-			return true;
-
-		// "why foo war?"
-		// chop off the "war" at the end
-		char aWhy[128];
-		str_copy(aWhy, pWhy, sizeof(aWhy));
-
-		int CutOffWar = -1;
-		if((CutOffWar = LangParser().StrFindIndex(aWhy, " war")) != -1)
-			aWhy[CutOffWar] = '\0';
-		else if((CutOffWar = LangParser().StrFindIndex(aWhy, " kill")) != -1)
-			aWhy[CutOffWar] = '\0';
-
-		// trim
-		int trim = 0;
-		while(aWhy[trim] == ' ')
-			trim++;
-
-		if(CutOffWar != -1)
-			if(WhyWar(aWhy + trim))
-				return true;
-
-		char aStripped[128];
-		StripSpacesAndPunctuationAndOwnName(pKill, aStripped, sizeof(aStripped));
-		if(!IsEmptyStr(aStripped))
-		{
-			str_format(m_pResponse, m_SizeOfResponse, "%s: '%s' is not on my warlist. axaxaxax", m_pMessageAuthor, aStripped);
-			return true;
-		}
-	}
+	// check war for others "is war name?"
+	if(IsWarName())
+		return true;
 	// why? (check war for self)
-	if(LangParser().IsQuestionWhy(m_pMessage) || (str_find(m_pMessage, "?") && MsgLen < NameLen + 4) ||
-		((str_find(m_pMessage, "stop") || str_find_nocase(m_pMessage, "help")) && (ChatHelper()->GameClient()->m_WarList.IsWarlist(m_pMessageAuthor) || ChatHelper()->GameClient()->m_WarList.IsTraitorlist(m_pMessageAuthor))))
-	{
-		char aWarReason[128];
-		if(ChatHelper()->GameClient()->m_WarList.IsWarlist(m_pMessageAuthor) || ChatHelper()->GameClient()->m_WarList.IsTraitorlist(m_pMessageAuthor))
-		{
-			ChatHelper()->GameClient()->m_WarList.GetWarReason(m_pMessageAuthor, aWarReason, sizeof(aWarReason));
-			if(aWarReason[0])
-				str_format(m_pResponse, m_SizeOfResponse, "%s has war because: %s", m_pMessageAuthor, aWarReason);
-			else
-				str_format(m_pResponse, m_SizeOfResponse, "%s you are on my warlist.", m_pMessageAuthor);
-			return true;
-		}
-		else if(ChatHelper()->GameClient()->m_WarList.IsWarClanlist(m_pMessageAuthorClan))
-		{
-			str_format(m_pResponse, m_SizeOfResponse, "%s your clan is on my warlist.", m_pMessageAuthor);
-			return true;
-		}
-		for(int i = 0; i < MAX_CLIENTS; i++)
-		{
-			auto &Client = ChatHelper()->GameClient()->m_aClients[i];
-			if(!Client.m_Active)
-				continue;
-			if(str_comp(Client.m_aName, m_pMessageAuthor))
-				continue;
-
-			if(ChatHelper()->GameClient()->m_WarList.IsWarClanmate(i))
-			{
-				str_format(m_pResponse, m_SizeOfResponse, "%s i might kill you because i war member of your clan", m_pMessageAuthor);
-				return true;
-			}
-		}
-	}
+	if(WhyDoYouKillMe(NameLen, MsgLen))
+		return true;
 	// still check war for others but now different order
 	// also cover "name is war?" in addition to "is war name?"
-	char aStrippedMsg[256];
-	StripSpacesAndPunctuationAndOwnName(m_pMessage, aStrippedMsg, sizeof(aStrippedMsg));
-	int ChopEnding = IsWarCheckSuffix(aStrippedMsg);
-	bool Strict = false;
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " bad");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " good");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " friend");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " frien");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " frent");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " fren");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " frint");
-	}
-	if(!ChopEnding)
-	{
-		Strict = true;
-		ChopEnding = GetSuffixLen(aStrippedMsg, " enemy");
-	}
-
-	if(ChopEnding)
-	{
-		unsigned int Cut = str_length(aStrippedMsg) - ChopEnding;
-		if(Cut > 0 && Cut < sizeof(aStrippedMsg))
-		{
-			aStrippedMsg[str_length(aStrippedMsg) - ChopEnding] = '\0';
-			if(WhyWar(aStrippedMsg))
-				return true;
-			bool FoundName = false;
-			if(Strict)
-			{
-				for(const auto &Client : ChatHelper()->GameClient()->m_aClients)
-				{
-					if(!Client.m_Active)
-						continue;
-					if(!str_comp(Client.m_aName, aStrippedMsg))
-					{
-						if(ChatHelper()->GameClient()->m_WarList.IsTeamlist(Client.m_aName))
-						{
-							str_format(m_pResponse, m_SizeOfResponse, "%s: '%s' is on my friend list.", m_pMessageAuthor, Client.m_aName);
-							return true;
-						}
-						if(ChatHelper()->GameClient()->m_WarList.IsTeamClanlist(Client.m_aClan))
-						{
-							str_format(m_pResponse, m_SizeOfResponse, "%s: '%s's clan %s is on my friend list.", m_pMessageAuthor, Client.m_aName, Client.m_aClan);
-							return true;
-						}
-						FoundName = true;
-						break;
-					}
-				}
-			}
-			const char *pPrefixStripped = aStrippedMsg;
-			// second search strip some prefix
-			if(!FoundName)
-			{
-				Strict = false;
-				if(str_startswith_nocase(aStrippedMsg, " is "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length(" is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else if(str_startswith_nocase(aStrippedMsg, "is "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length("is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else if(str_startswith_nocase(aStrippedMsg, " do you "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length("is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else if(str_startswith_nocase(aStrippedMsg, "do you "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length("is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else if(str_startswith_nocase(aStrippedMsg, " on "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length("is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else if(str_startswith_nocase(aStrippedMsg, "on "))
-				{
-					pPrefixStripped = aStrippedMsg + str_length("is ");
-					if(WhyWar(pPrefixStripped))
-						return true;
-				}
-				else
-					Strict = true;
-			}
-			if(!Strict || FoundName)
-			{
-				str_format(m_pResponse, m_SizeOfResponse, "%s: '%s' is not on my warlist.", m_pMessageAuthor, pPrefixStripped);
-				return true;
-			}
-		}
-	}
+	if(NameIsWar())
+		return true;
 
 	// check all wars "who is on your warlist?"
-	const char *pWho = str_find_nocase(m_pMessage, "who ");
-	if(pWho)
-		pWho = pWho + str_length("who ");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "which")))
-			pWho = pWho + str_length("which");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "got any")))
-			pWho = pWho + str_length("got any");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "have any")))
-			pWho = pWho + str_length("have any");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "wer ")))
-			pWho = pWho + str_length("wer ");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "wie viele")))
-			pWho = pWho + str_length("wie viele");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, " how ")))
-			pWho = pWho + str_length(" how ");
-	if(!pWho)
-		pWho = str_startswith_nocase(m_pMessage, "how ");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "howm"))) // howmani howmany howm any
-			pWho = pWho + str_length("howm");
-	if(!pWho)
-		if((pWho = str_find_nocase(m_pMessage, "u got")))
-			pWho = pWho + str_length("u got");
-	if(pWho)
-	{
-		const char *pEnemy = NULL;
-		pEnemy = str_find_nocase(pWho, "enemi");
-		if(pEnemy)
-			pEnemy = pEnemy + str_length("enemi");
-		else if((pEnemy = str_find_nocase(pWho, "enemy")))
-			pEnemy = pEnemy + str_length("enemy");
-		else if((pEnemy = str_find_nocase(pWho, "gegner")))
-			pEnemy = pEnemy + str_length("gegner");
-		else if((pEnemy = str_find_nocase(pWho, "against")))
-			pEnemy = pEnemy + str_length("against");
-		else if((pEnemy = str_find_nocase(pWho, "opponent")))
-			pEnemy = pEnemy + str_length("opponent");
-		else if((pEnemy = str_find_nocase(pWho, "raid")))
-			pEnemy = pEnemy + str_length("raid");
-		else if((pEnemy = str_find_nocase(pWho, "block")))
-			pEnemy = pEnemy + str_length("block");
-		else if((pEnemy = str_find_nocase(pWho, "kill")))
-			pEnemy = pEnemy + str_length("kill");
-		else if((pEnemy = str_find_nocase(pWho, "spaik")))
-			pEnemy = pEnemy + str_length("spaik");
-		else if((pEnemy = str_find_nocase(pWho, "spike")))
-			pEnemy = pEnemy + str_length("spike");
-		else if((pEnemy = str_find_nocase(pWho, "bad")))
-			pEnemy = pEnemy + str_length("bad");
-		else if((pEnemy = str_find_nocase(pWho, "war")))
-			pEnemy = pEnemy + str_length("war");
-		if(pEnemy)
-		{
-			char aEnemyList[256]; // 255 max msg len
-			aEnemyList[0] = '\0';
-			int NumEnemies = 0;
-			for(auto &Client : ChatHelper()->GameClient()->m_aClients)
-			{
-				if(!Client.m_Active)
-					continue;
-				if(!ChatHelper()->GameClient()->m_WarList.IsWarlist(Client.m_aName) &&
-					!ChatHelper()->GameClient()->m_WarList.IsWarClanlist(Client.m_aClan))
-					continue;
-
-				NumEnemies++;
-				if(aEnemyList[0])
-				{
-					str_append(aEnemyList, ", ", sizeof(aEnemyList));
-				}
-				str_append(aEnemyList, Client.m_aName, sizeof(aEnemyList));
-			}
-			if(NumEnemies)
-				str_format(m_pResponse, m_SizeOfResponse, "%s %d of my %d enemies are online: %s", m_pMessageAuthor, NumEnemies, ChatHelper()->GameClient()->m_WarList.NumEnemies(), aEnemyList);
-			else
-				str_format(m_pResponse, m_SizeOfResponse, "%s currently 0 of my %d enemies are connected", m_pMessageAuthor, ChatHelper()->GameClient()->m_WarList.NumEnemies());
-			return true;
-		}
-		const char *pFriend = NULL;
-		pFriend = str_find_nocase(pWho, "team");
-		if(pFriend)
-			pFriend = pFriend + str_length("team");
-		else if((pFriend = str_find_nocase(pWho, "friend")))
-			pFriend = pFriend + str_length("friend");
-		else if((pFriend = str_find_nocase(pWho, "frint")))
-			pFriend = pFriend + str_length("frint");
-		else if((pFriend = str_find_nocase(pWho, "peace")))
-			pFriend = pFriend + str_length("peace");
-		else if((pFriend = str_find_nocase(pWho, "like")))
-			pFriend = pFriend + str_length("like");
-		else if((pFriend = str_find_nocase(pWho, "good")))
-			pFriend = pFriend + str_length("good");
-		else if((pFriend = str_find_nocase(pWho, "help")))
-			pFriend = pFriend + str_length("help");
-		else if((pFriend = str_find_nocase(pWho, "side")))
-			pFriend = pFriend + str_length("side");
-		else if((pFriend = str_find_nocase(pWho, "with u")))
-			pFriend = pFriend + str_length("with u");
-		else if((pFriend = str_find_nocase(pWho, "with y")))
-			pFriend = pFriend + str_length("with y");
-		else if((pFriend = str_find_nocase(pWho, "u with")))
-			pFriend = pFriend + str_length("u with");
-		if(pFriend)
-		{
-			char aFriendList[256]; // 255 max msg len
-			aFriendList[0] = '\0';
-			int NumFriends = 0;
-			for(auto &Client : ChatHelper()->GameClient()->m_aClients)
-			{
-				if(!Client.m_Active)
-					continue;
-				if(!ChatHelper()->GameClient()->m_WarList.IsTeamlist(Client.m_aName) &&
-					!ChatHelper()->GameClient()->m_WarList.IsTeamClanlist(Client.m_aClan))
-					continue;
-
-				NumFriends++;
-				if(aFriendList[0])
-				{
-					str_append(aFriendList, ", ", sizeof(aFriendList));
-				}
-				str_append(aFriendList, Client.m_aName, sizeof(aFriendList));
-			}
-			if(NumFriends)
-				str_format(m_pResponse, m_SizeOfResponse, "%s %d of my %d friends are online: %s", m_pMessageAuthor, NumFriends, ChatHelper()->GameClient()->m_WarList.NumTeam(), aFriendList);
-			else
-				str_format(m_pResponse, m_SizeOfResponse, "%s currently 0 of my %d friends are connected", m_pMessageAuthor, ChatHelper()->GameClient()->m_WarList.NumTeam());
-			return true;
-		}
-	}
+	if(ListWars())
+		return true;
 
 	// intentionally check for being on warlist
 	// also expecting an no if not
-	if(str_find_nocase(m_pMessage, "?") ||
-		str_find_nocase(m_pMessage, "do you") ||
-		str_find_nocase(m_pMessage, "are we") ||
-		str_find_nocase(m_pMessage, "u war") ||
-		str_find_nocase(m_pMessage, "war me?") ||
-		str_find_nocase(m_pMessage, "am i") ||
-		str_find_nocase(m_pMessage, "is i") ||
-		str_find_nocase(m_pMessage, "im your") ||
-		str_find_nocase(m_pMessage, "what") ||
-		str_find_nocase(m_pMessage, "show") ||
-		str_find_nocase(m_pMessage, "wat"))
-	{
-		if(str_find_nocase(m_pMessage, "me enem") || str_find_nocase(m_pMessage, "i enem") || str_find_nocase(m_pMessage, "me is enem") || str_find_nocase(m_pMessage, "i is enem") ||
-			str_find_nocase(m_pMessage, "i u enem") || str_find_nocase(m_pMessage, "i ur enem") || str_find_nocase(m_pMessage, "i your enem") ||
-			str_find_nocase(m_pMessage, "im u enem") || str_find_nocase(m_pMessage, "im ur enem") || str_find_nocase(m_pMessage, "im your enem") ||
-			str_find_nocase(m_pMessage, "i am u enem") || str_find_nocase(m_pMessage, "i am ur enem") || str_find_nocase(m_pMessage, "i am your enem") ||
-			str_find_nocase(m_pMessage, "me war") || str_find_nocase(m_pMessage, "i war") || str_find_nocase(m_pMessage, "me is war") || str_find_nocase(m_pMessage, "i is war") ||
-			str_find_nocase(m_pMessage, "peace") ||
-			(str_find_nocase(m_pMessage, "war me?") && MsgLen < NameLen + 15) ||
-			str_find_nocase(m_pMessage, "u war me") ||
-			str_find_nocase(m_pMessage, "me friend") || str_find_nocase(m_pMessage, "i friend") || str_find_nocase(m_pMessage, "me is friend") || str_find_nocase(m_pMessage, "i is friend") ||
-			str_find_nocase(m_pMessage, "me frint") || str_find_nocase(m_pMessage, "i frint") || str_find_nocase(m_pMessage, "me is frint") || str_find_nocase(m_pMessage, "i is frint") ||
-			str_find_nocase(m_pMessage, "are we in war") || str_find_nocase(m_pMessage, "we war") || str_find_nocase(m_pMessage, "we peace") || str_find_nocase(m_pMessage, "we good") ||
-			str_find_nocase(m_pMessage, "i enem") || str_find_nocase(m_pMessage, "i peace") || str_find_nocase(m_pMessage, "i frien") || str_find_nocase(m_pMessage, "i frin") ||
-			str_find_nocase(m_pMessage, "me enem") || str_find_nocase(m_pMessage, "me peace") || str_find_nocase(m_pMessage, "me frien") || str_find_nocase(m_pMessage, "me frin") ||
-			str_find_nocase(m_pMessage, "colo") || str_find_nocase(m_pMessage, "cole") || str_find_nocase(m_pMessage, "collo") || str_find_nocase(m_pMessage, "colla") || str_find_nocase(m_pMessage, "cola") ||
-			str_find_nocase(m_pMessage, "red") || str_find_nocase(m_pMessage, "green") || str_find_nocase(m_pMessage, "orange") || str_find_nocase(m_pMessage, "black") || str_find_nocase(m_pMessage, "reason"))
-		{
-			char aWarReason[128];
-			if(ChatHelper()->GameClient()->m_WarList.IsWarlist(m_pMessageAuthor) || ChatHelper()->GameClient()->m_WarList.IsTraitorlist(m_pMessageAuthor))
-			{
-				ChatHelper()->GameClient()->m_WarList.GetWarReason(m_pMessageAuthor, aWarReason, sizeof(aWarReason));
-				if(aWarReason[0])
-					str_format(m_pResponse, m_SizeOfResponse, "%s you have war because: %s", m_pMessageAuthor, aWarReason);
-				else
-					str_format(m_pResponse, m_SizeOfResponse, "%s you are on my warlist.", m_pMessageAuthor);
-				return true;
-			}
-			else if(ChatHelper()->GameClient()->m_WarList.IsWarClanlist(m_pMessageAuthorClan))
-			{
-				str_format(m_pResponse, m_SizeOfResponse, "%s your clan is on my warlist.", m_pMessageAuthor);
-				return true;
-			}
-			for(int i = 0; i < MAX_CLIENTS; i++)
-			{
-				auto &Client = ChatHelper()->GameClient()->m_aClients[i];
-				if(!Client.m_Active)
-					continue;
-				if(str_comp(Client.m_aName, m_pMessageAuthor))
-					continue;
-
-				if(ChatHelper()->GameClient()->m_WarList.IsWarClanmate(i))
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i might kill you because i war member of your clan", m_pMessageAuthor);
-					return true;
-				}
-			}
-			str_format(m_pResponse, m_SizeOfResponse, "%s your name is neither on my friend nor enemy list.", m_pMessageAuthor);
-			return true;
-		}
-	}
+	if(DoYouWarMe(NameLen, MsgLen))
+		return true;
 
 	// where are you
-	if(str_find_nocase(m_pMessage, "where are you") || str_find_nocase(m_pMessage, "where r u") || (str_find_nocase(m_pMessage, "where r yo") && !str_find_nocase(m_pMessage, "wo bist")))
-	{
-		for(auto &Client : ChatHelper()->GameClient()->m_aClients)
-		{
-			if(!Client.m_Active)
-				continue;
-			if(str_comp(Client.m_aName, m_pMessageAuthor))
-				continue;
-
-			// TODO: dont get current dummy but the pinged dummy
-			CCharacter *pChar = ChatHelper()->GameClient()->m_GameWorld.GetCharacterByID(ChatHelper()->GameClient()->m_LocalIDs[g_Config.m_ClDummy]);
-			if(!pChar)
-				continue;
-			vec2 Self = pChar->m_Pos;
-			vec2 Other = Client.m_RenderPos;
-			float distY = abs(Self.y - Other.y);
-			float distX = abs(Self.x - Other.x);
-			if(distX < 5 * 32 && distY < 5 * 32)
-			{
-				if(distX > distY)
-				{
-					if(Self.x > Other.x)
-					{
-						str_format(m_pResponse, m_SizeOfResponse, "%s i am literally next to you (on your right)", m_pMessageAuthor);
-						return true;
-					}
-					else
-					{
-						str_format(m_pResponse, m_SizeOfResponse, "%s i am literally next to you (on your left)", m_pMessageAuthor);
-						return true;
-					}
-				}
-				else
-				{
-					if(Self.y > Other.y)
-					{
-						str_format(m_pResponse, m_SizeOfResponse, "%s i am literally next to you (below you)", m_pMessageAuthor);
-						return true;
-					}
-					else
-					{
-						str_format(m_pResponse, m_SizeOfResponse, "%s i am literally next to you (above you)", m_pMessageAuthor);
-						return true;
-					}
-				}
-			}
-			else if(distY > distX * 8)
-			{
-				if(Self.y > Other.y)
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s below you (%d tiles)", m_pMessageAuthor, (int)(distY / 32));
-					return true;
-				}
-				else
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s above you (%d tiles)", m_pMessageAuthor, (int)(distY / 32));
-					return true;
-				}
-			}
-			else if(Self.x > Other.x)
-			{
-				if(distY < 15 * 32)
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your right (%d tiles away)", m_pMessageAuthor, (int)(distX / 32));
-					return true;
-				}
-				else if(Self.y > Other.y)
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your right (%d tiles away) and %d tiles below you", m_pMessageAuthor, (int)(distX / 32), (int)(distY / 32));
-					return true;
-				}
-				else
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your right (%d tiles away) and %d tiles above you", m_pMessageAuthor, (int)(distX / 32), (int)(distY / 32));
-					return true;
-				}
-			}
-			else
-			{
-				if(distY < 15 * 32)
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your left (%d tiles away)", m_pMessageAuthor, (int)(distX / 32));
-					return true;
-				}
-				else if(Self.y > Other.y)
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your left (%d tiles away) and %d tiles below you", m_pMessageAuthor, (int)(distX / 32), (int)(distY / 32));
-					return true;
-				}
-				else
-				{
-					str_format(m_pResponse, m_SizeOfResponse, "%s i am on your left (%d tiles away) and %d tiles above you", m_pMessageAuthor, (int)(distX / 32), (int)(distY / 32));
-					return true;
-				}
-			}
-		}
-		str_format(m_pResponse, m_SizeOfResponse, "%s no idea. Where are you?", m_pMessageAuthor);
+	if(Where())
 		return true;
-	}
 
 	// spec me
 	if(str_find_nocase(m_pMessage, "spec") || str_find_nocase(m_pMessage, "watch") || (str_find_nocase(m_pMessage, "look") && !str_find_nocase(m_pMessage, "looks")) || str_find_nocase(m_pMessage, "schau"))
@@ -949,33 +148,8 @@ bool CReplyToPing::Reply()
 			return true;
 		}
 	}
-	// small talk
-	if(str_find_nocase(m_pMessage, "how are you") ||
-		str_find_nocase(m_pMessage, "how r u") ||
-		str_find_nocase(m_pMessage, "how ru ") ||
-		str_find_nocase(m_pMessage, "how ru?") ||
-		str_find_nocase(m_pMessage, "how r you") ||
-		str_find_nocase(m_pMessage, "how are u") ||
-		str_find_nocase(m_pMessage, "how is it going") ||
-		str_find_nocase(m_pMessage, "ca va") ||
-		(str_find_nocase(m_pMessage, "как") && str_find_nocase(m_pMessage, "дела")))
-	{
-		str_format(m_pResponse, m_SizeOfResponse, "%s good, and you? :)", m_pMessageAuthor);
+	if(SmallTalk(NameLen, MsgLen))
 		return true;
-	}
-	if(str_find_nocase(m_pMessage, "wie gehts") || str_find_nocase(m_pMessage, "wie geht es") || str_find_nocase(m_pMessage, "was geht"))
-	{
-		str_format(m_pResponse, m_SizeOfResponse, "%s gut, und dir? :)", m_pMessageAuthor);
-		return true;
-	}
-	if(str_find_nocase(m_pMessage, "about you") || str_find_nocase(m_pMessage, "and you") || str_find_nocase(m_pMessage, "and u") ||
-		(str_find_nocase(m_pMessage, "u?") && MsgLen < NameLen + 5) ||
-		(str_find_nocase(m_pMessage, "wbu") && MsgLen < NameLen + 8) ||
-		(str_find_nocase(m_pMessage, "hbu") && MsgLen < NameLen + 8))
-	{
-		str_format(m_pResponse, m_SizeOfResponse, "%s good", m_pMessageAuthor);
-		return true;
-	}
 	// chillerbot-ux features
 	if(LangParser().IsQuestionHow(m_pMessage))
 	{
